CASMCMS-8557: Add convenience CLI options for BOS v2 components updates
